Finished up the slave portion of the revised twi driver

+++ runtime/ats/basics.h
@@ -99,7 +99,7 @@ typedef void ats_void_type ;
 typedef char ats_char_type ;
 typedef char ats_schar_type ;
 typedef unsigned char ats_uchar_type ;
-typedef int ats_bool_type ;
+typedef uint8_t ats_bool_type ;
 typedef double ats_double_type ;
 typedef size_t ats_size_type ;
